+++ app/src/main/java/com/esatgozcu/animationexamples/helper/Constants.kt
@@ -1,6 +1,6 @@
 package com.esatgozcu.animationexamples.helper
 
 object ScreenSize {
-    var screensWidth : Int = 0
-    var screenHeight: Int = 0
+    var screensWidth : Double = 0.0
+    var screenHeight: Double = 0.0
 }
